requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
alembic
numpy
pandas
scikit-learn
torch
torchvision
opencv-python
nltk
spacy
pillow
matplotlib
seaborn
python-dotenv
pytest
streamlit
plotly
httpx
jinja2 